fix(02.2-04): merge /settings/account into /settings; fix OAuth client; / → /about
Continued UAT cleanup found during the live aeza deploy. Three loosely
related fixes that all surfaced from real navigation testing:

#1 — Account UI was split across two routes nobody could find
/settings had an "Account" block (name + email + sign-out) but no link
to /settings/account where Export + Delete lived. Operators following
the runbook had to know the URL by hand. Merged the Export + Delete UI
into /settings's existing Account block as sub-sections (separator
border + h3 sub-headings) so all account-actions are in one cohesive
plate. Old /settings/account route now 301-redirects to /settings so
external bookmarks don't break.

#2 — Better Auth client was calling the wrong endpoint
auth-client.ts created a plain createAuthClient() with no plugins.
signIn.social({ provider: "google" }) tried to POST /api/auth/sign-in/
social/google — but our server uses the genericOAuth plugin (see
src/lib/auth.ts), which exposes /api/auth/sign-in/oauth2 instead. Local
dev login 500'd; production happened to work for an unrelated reason
(the prod-mounted handler tolerated the mismatched path).

Fix: register genericOAuthClient() in createAuthClient's plugins. That
adds signIn.oauth2({ providerId, callbackURL }) which posts to the
correct endpoint. Updated /login and /about (new CTA) to call
signIn.oauth2 instead of signIn.social.

#3 — Anonymous landing redirect: / → /about
Root / for anonymous users used to render a thin Phase 1 placeholder
("welcome + sign-in CTA"). UAT feedback: nobody scrolls a placeholder;
nobody knows /about exists. /about is now a rich landing page (5
sections + "Continue with Google" CTA in hero + legal-footer). Make
/about the canonical anonymous entry point: / → /about (303 redirect)
for unauthenticated requests; signed-in still go to /feed.

GitHub URL fix sneak-in: messages/en.json + docs/deploy/install.md +
src/routes/about/+page.svelte all referenced
github.com/d954mas/neotolis-diary (the docker package name) instead of
the actual repo at github.com/d954mas/neotolis-game-promotion-diary.
Fixed all four references.

138 unit tests pass; typecheck 0 errors. Local dev verified end-to-end
(mock OAuth via dev-mock-oauth.mjs, postgres via `docker run -p
5432:5432`); prod UAT pending PR merge.
