+++ messages/en.json
@@ -298,7 +298,7 @@
   "privacy_title": "Privacy Policy",
   "privacy_last_updated": "Last updated: {date}",
   "privacy_section_who_we_are_title": "Who we are",
-  "privacy_section_who_we_are_body": "Neotolis Game Promotion Diary is operated by an indie game developer. The service is open-source and MIT-licensed; the source code lives at https://github.com/d954mas/neotolis-diary. For all data-protection questions, contact {supportEmail}.",
+  "privacy_section_who_we_are_body": "Neotolis Game Promotion Diary is operated by an indie game developer. The service is open-source and MIT-licensed; the source code lives at https://github.com/d954mas/neotolis-game-promotion-diary. For all data-protection questions, contact {supportEmail}.",
   "privacy_section_data_collected_title": "What data we collect",
   "privacy_section_data_collected_body": "Identity: your Google email address and your unique Google account identifier. Usage: every API request's IP address and user-agent are recorded in the per-user audit log. User content: game cards, registered data sources, events (videos, posts, conferences), notes, tags — exactly the content you enter. OAuth tokens (Google access + refresh tokens) AND per-user API keys are envelope-encrypted at rest with a per-row DEK wrapped by a KEK loaded from the environment.",
   "privacy_section_lawful_basis_title": "Lawful basis (GDPR Art. 6)",
@@ -337,7 +337,7 @@
   "terms_section_early_access_title": "Early access disclaimer",
   "terms_section_early_access_body": "The service is in early access. Data preservation is best-effort. Major schema changes may force a re-import. The service may be paused with 14 days' notice.",
   "terms_section_open_source_title": "Open source",
-  "terms_section_open_source_body": "The code is MIT-licensed at https://github.com/d954mas/neotolis-diary. The hosted SaaS instance is offered as-is.",
+  "terms_section_open_source_body": "The code is MIT-licensed at https://github.com/d954mas/neotolis-game-promotion-diary. The hosted SaaS instance is offered as-is.",
   "terms_section_warranty_title": "No warranty",
   "terms_section_warranty_body": "The service is provided AS IS, without warranty of any kind. The operator is not liable for indirect, special, or consequential damages.",
   "terms_section_termination_title": "Termination",
