+++ src/routes/about/+page.server.ts
@@ -7,9 +7,13 @@ import { env } from "$lib/server/config/env.js";
 // the canonical-instance link — empty in self-host / CI / dev so the
 // link only appears on the canonical author-instance.
 
-export const load: PageServerLoad = () => {
+export const load: PageServerLoad = async ({ parent }) => {
+  // Pull user from layout via parent() so the page can hide/show the
+  // "Sign in with Google" CTA based on auth state.
+  const parentData = await parent();
   return {
     supportEmail: env.SUPPORT_EMAIL,
     domain: env.DOMAIN,
+    user: parentData.user,
   };
 };
